upadte

+++ menus/pgtrak/main.sh
@@ -18,15 +18,22 @@
 export NCURSES_NO_UTF8_ACS=1
 
 ### Pull the API Keys
-info=$( cat /opt/appdata/radarr/config.xml 1>/dev/null 2>&1 ) 
+file="/opt/appdata/radarr/config.xml"
+if [ -e "$file" ]
+then
+info=$( cat /opt/appdata/radarr/config.xml ) 
 info=${info#*<ApiKey>} 1>/dev/null 2>&1
 info1=$( echo ${info:0:32} ) 1>/dev/null 2>&1
 echo "$info1" > /var/plexguide/api.radarr
 
-info=$( cat /opt/appdata/sonarr/config.xml 1>/dev/null 2>&1 ) 
-info=${info#*<ApiKey>} 1>/dev/null 2>&1
-info2=$( echo ${info:0:32} ) 1>/dev/null 2>&1
-echo "$info2" > /var/plexguide/api.sonarr
+file="/opt/appdata/sonarr/config.xml"
+if [ -e "$file" ]
+then
+    info=$( cat /opt/appdata/sonarr/config.xml ) 
+    info=${info#*<ApiKey>} 1>/dev/null 2>&1
+    info2=$( echo ${info:0:32} ) 1>/dev/null 2>&1
+    echo "$info2" > /var/plexguide/api.sonarr
+fi
 
 ### If Both are Blank, which means neither deployed, warn USER!
 if [ "$info2" == "$info1" ]
